sima_httprunner3/requirements.txt
allure-pytest==2.8.6
httprunner==3.1.5
requests_toolbelt==0.9.1
filetype==1.0.7